Module29.2: moving 3D player including animations
