+++ 3D-EBAC-Action/Assets/Scenes/SCN_Art_3D.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fcb271660b922ca185cd2f7c1a1a4c0c68bf28fd43ada149185abd571b249310
-size 511347
+oid sha256:dc56290849827de75ab5fb9894f982629e37a12057462b66c31c7223a6d3da01
+size 516597
