+++ 3D-EBAC-Action/Assets/Scripts/Player/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b54424d6a2a0e6549acf8375076975cc
+guid: c0ae5d5252d3147dda193ddff43489c9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
